requirements.txt
streamlit==1.32.0
pandas==2.1.0
numpy==1.24.3
plotly==5.18.0
pydeck==0.8.0
pyyaml==6.0